refactor: `apply_interrogative_upspeak()` 移植
