+++ test/test_synthesis_engine_base.py
@@ -6,7 +6,7 @@ import numpy
 
 from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
 from voicevox_engine.tts_pipeline import TTSEngine
-from voicevox_engine.tts_pipeline.tts_engine_base import apply_interrogative_upspeak
+from voicevox_engine.tts_pipeline.tts_engine import apply_interrogative_upspeak
 
 
 def yukarin_s_mock(length: int, phoneme_list: numpy.ndarray, style_id: numpy.ndarray):
